[schroot] setup: Copy NSS databases using getent
20nssdatabases, nssdatabases-defaults: New script to copy the
databases listed in nssdatabases-defaults into the chroot.
Databases from /etc are no longer copied with 20copyfiles (with
the exception of gshadow, which does not appear to have NSS support).
As a result, the host NSS setup will result in functional file-based
database support inside the chroot.

Installations with very large database files may wish to use
other NSS configuration inside the chroot in order to avoid
performance issues duplicating the databases on chroot setup,
but for most users this setup will be sufficiently flexible.