[sbuild::chroot_block_device] Chain up to chroot in copy constructor
Since chroot is pure virtual, it must be chained up to in the
copy constructor of the most derived class or else it gets
default constructed.  This is needed for lvm-snapshot source
chroot construction (which worked previously due to block_device
being a direct base).

Long-term, the virtual public chroot inheritance needs removing.