sbuild::chroot_facet_session_clonable: Only allow session creator to access session
Previously, any user who had access to a chroot automatically gained
access to any sessions created from the chroot.  This has now changed.
Only the user creating the session will be granted access, in order
to improve the security of sessions.

The chroot clone_session method takes two additional arguments: the
user who will own the session, and whether or not the user is to
have root access.  This is used by the session code, and eventually
passed to the session clone_session_setup method which will update
the properties to match what was requested.