sbuild: Add reentrant passwd and group wrappers
In order to make the library thread-safe, use the reentrant
getpw*_r and getgr*_r functions in place of the older unsafe
functions which use static buffers.  The passwd and group
structures are inherited by C++ classes of the same name which
allow handling of dynamic memory allocation while at the same
time allowing use of RAII and passing to C functions.