build: boost_filesystem depends upon boost_system
Newer compilers (GCC 4.5) and linkers (gold) require that
all required symbols are linked against directly, rather than
allowing indirect linkage to satisfy the dependency.  This
means that if you use the libboost_filesystem headers and
link with -lboost_filesystem, you also need to add -lboost_system.
The library has the appropriate DT_NEEDED entry, but this is due
to exposing internal use of boost_system in the boost_filesystem
headers, so our code is in fact dependent even though we are
unaware of it.

Hopefully boost will start using pkg-config in the future so we
don't need to be aware of these internal details and worry about
breakage when our assumptions about the internal details change
in the future.