setup: 10mount: Use flock to prevent race reading /proc/mounts
See Debian #593516 for a full description of this bug.

Comment added to Greg Price's patch by Roger Leigh:

Note that flock is used here to prevent races reading /proc/mounts,
which on current (Linux 2.6.32) kernels is racy.  If other processes
are mounting or unmounting filesystems as we read it, we can miss
mount entries due to the file changing as we read it.  This needs
fixing in the kernel, but an exclusive lock surrounding the
schroot-listmounts invocation is a partial fix.  This prevents racing
when multiple schroot processes are running.  Note that this does not
prevent the problem when programs other than schroot mount and unmount
filesystems (since they don't create the lock).