15killprocs: Device and inode do not uniquely identify a chroot
15killprocs switched to comparing device/inode numbers to kill
processes inside a chroot.  Unfortunately, when multiple sessions
are run inside a non-cloned chroot, the sessions all share a the
same device/inode as the chroot root directory, resulting in all
processes in all the sessions being killed when one of the sessions
is ended.

Retain the device/inode check for efficiency, but additionally
check the paths when a match is found to avoid killing processes
in other sessions.