15killprocs: Warn about race condition when killing processes
Invoking /bin/kill on processes fails if the process already
quit by itself while the schroot shutdown runs. Because of "set -e"
this causes the whole shutdown to fail.

I have seen this occasionally in a nightly testing setup involving
a D-Bus session: when the D-Bus daemon stops, all other processes
depending on it also shut down automatically, with or without being
killed.