build: Move all path specification to global.mk
It's not possible to evaluate directory variables at configure
time, even with eval.  Move all preprocessor definitions to
DEFS in global.mk.

Note that this also required changing the substitution mechanism
in the manual pages, which have been changed to source a
generated config.man fragment, which is sourced using the .so
directive, and the manual page generated using soelim(1).
The boilerplate for the authors and copyright sections has also
been split out at the same time.