schroot-mount: Canonicalise symlinks to avoid host mounts
If a symlink inside the chroot is absolute, and this is used
as part of the path in a mountpoint in the fstab file, then
this currently gets mounted (incorrectly) on the host.  Use
realpath(3) to canonicalise the path, and then add the chroot
path back to it if missing, to ensure that symlinks can't be
used to accidentally mount filesystems on the host.