build: Correct regex checks
cmake checks for working <regex> were incomplete since the existing
logic had not had a working GCC to build against until now.  These
are fixed with this commit.

While GCC 4.9 now supports <regex>, the support appears to still be
broken.  See
  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=61059

Extend the tests to check for this specific failure case to ensure
boost::regex is still used with GCC 4.9.