etc/setup.d/20copyfiles: Replace dangling symlink during cp. #1
Add --remove-destination to the cp calls in etc/setup.d/20copyfiles to
fix problems with dangling symlinks.

Error message from cp is:
  cp: not writing through dangling symlink ‘asymlinkfile’

This is useful since /etc/resolv.conf may be a dangling symlink ( which
happens if resolvconf is installed in the schroot, or if the schroot
uses systemd and /etc/resolv.conf points to the non-existant
/run/systemd/resolve/resolv.conf ).

Adding --remove-destination fixes this. (cp -f doesn't fix it.)